feat(tinny-helper): make `stringToIpfsHash` a tinny helper
